evering_ipc::shm

Struct ShmBox

pub struct ShmBox<T: ?Sized>(/* private fields */);

Implementations

impl<T: ?Sized> ShmBox<T>

pub fn as_shm(this: &Self) -> ShmToken<T>

pub fn into_raw(self) -> NonNull<T>

pub unsafe fn from_raw(ptr: NonNull<T>) -> Self

Safety

The given ptr must be a valid pointer managed by Allocator.

impl<T> ShmBox<T>

pub fn new(val: T) -> Self

pub fn into_uninit(self) -> ShmBox<MaybeUninit<T>>

impl<T> ShmBox<MaybeUninit<T>>

pub fn new_uninit() -> Self

pub unsafe fn assume_init(self) -> ShmBox<T>

Safety

The underlying value must be initialized properly. See MaybeUninit::assume_init for more information.

impl<T> ShmBox<[T]>

pub fn new_slice_copied(src: &[T]) -> Self
where T: Copy,

pub fn new_slice_filled(val: T, n: usize) -> Self
where T: Copy,

pub fn into_uninit(self) -> ShmBox<[MaybeUninit<T>]>

impl<T> ShmBox<[MaybeUninit<T>]>

pub fn new_slice_uninit(n: usize) -> Self

pub unsafe fn assume_init(self) -> ShmBox<[T]>

Safety

Each element in the underlying slice must be initialized properly. See MaybeUninit::assume_init for more information.