evering_ipc::shm

Struct ShmHeader

pub struct ShmHeader<A = Sqe, B = Rqe, Ext = ()> { /* private fields */ }

ShmHeader contains necessary metadata of a shared memory region.

Memory layout of the entire shared memory is illustrated as below,

.-----------------------------------------------------------------------------.
|                   |               |                   |                     |
| [1] uring offsets | [2] allocator | [3] uring buffers | [4] free memory ... |
| ^                 |               |                   |                   ^ |
'-|-------------------------------------------------------------------------|-'
  '-- start of the shared memory (page aligned)                             |
                                                 end of the shared memory --'

Uring offsets are used to build [RawUring].

Each shared memory region comes with one single-thread allocator. Typically, it will be taken by the client after initialization.

The submitted requests and responses are stored in these buffers.

The rest of the shared memory are managed by the allocator. ShmBox provides similar APIS to Box, but it is allocated and deallocated by the shared memory Allocator instead of the global allocator.

Implementations

impl<A, B, Ext> ShmHeader<A, B, Ext>

pub unsafe fn create( fd: BorrowedFd<'_>, size: usize, header: UringHeader<Ext>, ) -> Result<NonNull<Self>>

Safety

The given fd must be valid for the remaining lifetime of the running program.

pub unsafe fn open(fd: BorrowedFd<'_>, size: usize) -> Result<NonNull<Self>>

Safety

The given fd must be valid for the remaining lifetime of the running program.

pub unsafe fn close(ptr: NonNull<Self>, size: usize) -> Result<()>

Safety

The supplied ptr and size must match the previous mmap call.

pub fn build_raw_uring(&self) -> RawUring<A, B, Ext>

pub fn get_allocator(&self) -> &Allocator

pub fn get_shm<T: ?Sized>(&self, ptr: NonNull<T>) -> ShmToken<T>

pub fn get_ptr<T: ?Sized>(&self, shm: ShmToken<T>) -> NonNull<T>

Safety

The given shm must belong to this memory region.